15.2.Base Test Class full of Goodies{Encoding the User Password}
Encoding the User Password

Ok, this feels good. What else can we do? The weirdest thing now is probably that we're passing this long encoded password. It's not obvious that this is an encoded version of the password foo and... it's annoying! Heck, I had to undo 2 minutes of work earlier so I could copy it!

Let's do this properly. Replace that huge, encoded password string with just foo.

Now, inside the base test class, remove the $password variable and replace it with $encoded = and... hmm. We need to get the service out of the container that's responsible for encoding passwords. We can get that with self::$container->get('security.password_encoder'). We also could have used UserPasswordEncoderInterface::class as the service id - that's the type-hint we use normally for autowiring. Now say ->encodePassword() and pass the$user object and then the plain text password: . Finish this with $user->setPassword($encoded).

Beautiful!

And this point... I'm happy! Heck, I'm thrilled! But... I do have one more shortcut idea. It'll be pretty common for us to want to create a user and then log in immediately. Let's make that easy! Add a protected function createUserAndLogIn()... which needs the same three arguments as the function above: $client,$email and $password... and we'll return the User. Inside say $user = $this->createUser() with $email and$password, then $this->logIn() with $client, $email, $password. At the bottom, return $user.

Nice! Now we can shorten things a little bit more: $this->createUserAndLogIn().

Let's try it! Run:

php bin/phpunit
All green