16.3.ACL: Only Owners can PUT a CheeseListing
Let's get our test back to green.

The access_control option - which will probably be renamed to security in API Platform 2.5 - allows you to write an "expression" inside using Symfony's expression language. This is_granted() thing is a function that's available in that, sort of, Twig-like expression language.

We can make this expression more interesting by saying and to add more logic. API Platform gives us a few variables to work with inside the expression, including one that represents the object we're working with on this operation... in other words, the CheeseListing object. That variable is called... object! Another is user, which is the currently-authenticated User or null if the user is anonymous.

Knowing that, we can say and object.getOwner() == user.

Yea... that's it! Try the test again and...

php bin/phpunit --filter=testUpdateCheeseListing
It passes! I told you the security part of this was going to be easy! Most of the work was the test, but I love that I can prove this works.