16.4.ACL: Only Owners can PUT a CheeseListing{access_control_message}
access_control_message

While we're here, there's one other related option called access_control_message. Set this to:

  only the creator can edit a cheese listing

... and make sure you have a comma after the previous line.

If you run the test... this makes no difference. But this option did just change the message the user sees. Check it out: after the 403 status code, var_dump() $client->getResponse()->getContent() and pass that false. Normally, if you call getContent() on an "error" response - a 400 or 500 level response - it throws an exception. This tells it not to, which will let us see that response's content. Try the test:
php bin/phpunit --filter=testUpdateCheeseListing
The hydra:title says "An error occurred" but the hydra:description says:

only the creator can edit a cheese listing.

So, the access_control_message is a nice way to improve the error your user sees. By the way, in API Platform 2.5, it'll probably be renamed to security_message.